Fix error on report saving
When a user update a report with project version blank,
an error raises caused by accessing method on null
object.

Known issue: when a user doesn't change project version
combo, even if it has any value, the value is always
null on the server side. Trying to figure out the
cause of this behavior.